Thanks to Rusty Russell: gets takes a size_t length.
git-svn-id: http://tangram.dnsalias.net/repos/pwlib-c/cfile@104 6bb8afd2-ee04-0410-8f1a-d3aebcb25db2

+++ cfile.c
@@ -318,7 +318,7 @@ bool cfeof(cfile *fp) {
  *     A pointer to the string thus read.
  */
 
-char *cfgets(cfile *fp, char *str, int len) {
+char *cfgets(cfile *fp, char *str, size_t len) {
     if (!fp || !fp->vptr) {
         errno = EINVAL;
         return NULL;
